requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apache-beam
attrs==25.3.0
avro-python3==1.10.2
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
crcmod==1.7
cryptography==44.0.2
curl_adapter==1.0.0.post3
curl_cffi==0.10.0
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
et_xmlfile==2.0.0
fastapi==0.115.12
fastavro==1.10.0
fasteners==0.19
finvizfinance==1.1.0
frozendict==2.4.6
frozenlist==1.6.0
google-api-core==2.24.2
google-apitools==0.5.31
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-bigquery-storage==2.31.0
google-cloud-bigtable==2.30.1
google-cloud-core==2.4.3
google-cloud-datastore==2.21.0
google-cloud-dlp==3.29.0
google-cloud-language==2.17.1
google-cloud-pubsub==2.29.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.17
google-cloud-resource-manager==1.14.2
google-cloud-spanner==3.54.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-genai==1.21.1
httpx
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.16.0
hdfs==2.7.3
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jiter==0.9.0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
lxml==5.4.0
monotonic==1.6
more-itertools==10.7.0
multidict==6.4.3
multitasking==0.0.11
numpy==2.2.5
oauth2client==4.1.3
objsize==0.7.1
openai==1.76.2
openbb-benzinga==1.4.1
openbb-core==1.4.7
openbb-finviz==1.3.0
openbb-fmp==1.4.2
openbb-multpl==1.1.0
openbb-yfinance==1.4.4
openbb-cftc
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.18.1
platformdirs==4.3.7
pluggy==1.5.0
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==16.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycurl==7.45.3
pydantic==2.11.4
pydantic_core==2.33.2
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.12.1
pykalman
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
ruff==0.7.4
scipy
SecretStorage==3.3.3
sendgrid==6.11.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
starlette==0.46.2
ta==0.11.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid7==0.1.0
uvicorn==0.34.2
vaderSentiment==3.3.2
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.59
zipp==3.21.0
zstandard==0.23.0
matplotlib
vix_utils
db-dtypes